+++ tests/testthat/test-async.R
@@ -287,3 +287,13 @@ test_that("async function returns invisibly (#46)", {
   })
   expect_invisible(fn())
 })
+
+test_that("async functions do not cause CMD check notes (#40)", {
+  skip_on_cran()
+  expect_silent(
+    invisible(compiler::cmpfun(
+      async(function() NULL),
+      options = list(suppressAll = FALSE)
+    ))
+  )
+})